Patch by: Xing Xue, Sameer.
With schedule dynamic, the XL compiler passes in different, non-overlapping ranges
per team to the dispatch_init runtime call.  Forcing the lower bound to 0 will cause
incorrect results for XL.  It works with Clang because it passes 0 for both teams.
Removing the highlighted statements from the runtime would make the runtime work
with both the XL compiler and Clang.